Move trace utils to common package
