+++ tests/unit/trace/test_utils.py
@@ -19,7 +19,7 @@ import unittest
 
 import mock
 
-from opencensus.trace import utils
+from opencensus.common import utils
 
 
 class TestUtils(unittest.TestCase):
@@ -37,7 +37,7 @@ class TestUtils(unittest.TestCase):
     def test__get_truncatable_str_length_exceeds(self):
         max_len = 5
         str_to_convert = 'length exceeded'
-        patch = mock.patch('opencensus.trace.utils.MAX_LENGTH', max_len)
+        patch = mock.patch('opencensus.common.utils.MAX_LENGTH', max_len)
 
         with patch:
             truncatable_str = utils._get_truncatable_str(str_to_convert)
